pyproject.toml
[project]
name = "agent-land"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.0.2",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "langchain-openai>=1.0.0",
    "typing-extensions>=4.0.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
